pyproject.toml
[project]
name = "latex-to-docx-beautify"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pypandoc-binary>=1.16.2",
]
